refactor delete piece modal
